resize numeric columns holding prices
Some currencies have 3 decimal places but we only had support for 2 and
at the same time allow more digits before the decimal points for crazy
currencies that are many K to the EUR. Cost of storage is identical for
Pg since cost is based on actual number stored. For MySQL increase in
cost is 3 bytes per value (one byte extra just for the extra DP).